split UI tests into a separate project
