+++ LSAnalyzer/README.md
@@ -2,11 +2,10 @@
 Test and debug!
 
 * split application core and UI in separate assemblies to possibly switch to AvaloniaUI
-* create an automated test suite for a wider range of published trustful coefficients from known data (PIRLS, PISA school level, ...) that conducts realistic workflows
 * refactoring of data table creation and management OR switch from BIFIEsurvey to a specialized R package for LSAnalyzer
 
 # Nice-to-have
-* actual list of weight variables with descriptions and optinal weight variables
+* actual list of weight variables with descriptions and optional weight variables
 
 # Would-be-nice-but-hard-to-implement
 * handle SPSS files in folders with non-ASCII characters
