+++ TestLSAnalyzerUI/_testData/dataProviders.json
@@ -1,0 +1,1 @@
+{"DataProviders":[{"$type":"dataverse","Id":2,"Name":"Test Dataverse","Url":"https://test.local/","ApiToken":"safafs324324f","HasErrors":false}]}
